tests/wpt/tests/html/webappapis/dynamic-markup-insertion/document-write/script_002.html

<!doctype html>
<title>document.write script executed synchronously</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var order = [];
t.step(function() {
  document.write("<script>t.step(function() {order.push(1);});<"+"/script>");
  order.push(2);
});
</script>
<script>
t.step(function() {
  order.push(3);
  assert_array_equals(order, [1,2,3]);
})
t.done();
</script>
<div id="log"></div>
